LPD-51363 Group field components
