+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/field_components/SingleSelectFieldComponents.tsx
@@ -5,18 +5,18 @@
 
 import React from 'react';
 
-import {Field} from '../utils/field';
-import MaxLengthInput from './MaxLengthInput';
+import {Field} from '../../utils/field';
+import PicklistPicker from '../PicklistPicker';
 
-export default function getLongTextFieldComponents(): {
+export default function getSingleSelectFieldComponents(): {
 	FirstSectionComponent?: React.FC<{field: Field}>;
 	SecondSectionComponent?: React.FC<{field: Field}>;
 } {
 	return {
-		SecondSectionComponent,
+		FirstSectionComponent,
 	};
 }
 
-function SecondSectionComponent({field}: {field: Field}) {
-	return <MaxLengthInput field={field} />;
+function FirstSectionComponent({field}: {field: Field}) {
+	return <PicklistPicker field={field} />;
 }
